Move build_tax_benefit_sytem to scripts
As it is reusable
