+++ openfisca_core/scripts/__init__.py
@@ -2,6 +2,7 @@
 
 import importlib
 import sys
+import importlib
 
 
 TAX_BENEFIT_SYSTEM_OPTIONS = {
@@ -33,6 +34,37 @@ def add_tax_benefit_system_arguments(parser):
             )
 
     return parser
+
+
+def build_tax_benefit_sytem(country_package, extensions, reforms):
+    if country_package:
+        try:
+            country_package = importlib.import_module(country_package)
+        except:
+            print('ERROR: `{}` does not seem to be a valid Openfisca country package.'.format(country_package))
+            sys.exit(1)
+    else:
+        country_package_name = detect_country_package()
+        country_package = importlib.import_module(country_package_name)
+
+    tax_benefit_system = country_package.CountryTaxBenefitSystem()
+
+    if extensions:
+        for extension in extensions:
+            tax_benefit_system.load_extension(extension)
+
+    if reforms:
+        for reform_path in reforms:
+            try:
+                [reform_package, reform_name] = reform_path.rsplit('.', 1)
+                reform_module = importlib.import_module(reform_package)
+                reform = getattr(reform_module, reform_name)
+                tax_benefit_system = reform(tax_benefit_system)
+            except:
+                print('ERROR: `{}` does not seem to be a valid Openfisca reform for `{}`.'.format(reform_path, country_package.__name__))
+                raise
+
+    return tax_benefit_system
 
 
 def detect_country_package():
